Addressed review comments.
